+++ src/main/java/cart/controller/dto/CartResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.controller.dto;
 
 public class CartResponse {
