+++ src/main/java/cart/controller/dto/ProductRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.controller.dto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.PositiveOrZero;
