+++ src/main/java/cart/controller/view/ProductViewController.java
@@ -3,7 +3,7 @@ package cart.controller.view;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import cart.dto.ProductResponse;
+import cart.controller.dto.ProductResponse;
 import cart.service.ProductSearchService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
