+++ src/main/java/cart/repository/RdsCartRepository.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import cart.controller.dto.CartResponse;
 import cart.dao.CartDao;
 import cart.dao.ProductDao;
 import cart.domain.cart.Cart;
 import cart.domain.product.Product;
 import cart.domain.user.Email;
-import cart.dto.CartResponse;
 import cart.entiy.CartEntity;
 import cart.entiy.ProductEntity;
 import org.springframework.stereotype.Repository;
