+++ src/test/java/cart/controller/api/ProductCreateControllerTest.java
@@ -8,10 +8,10 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import cart.controller.AbstractProductControllerTest;
+import cart.controller.dto.ProductRequest;
+import cart.controller.dto.ProductResponse;
 import cart.domain.product.Product;
 import cart.domain.product.ProductFixture;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
 import cart.dto.RequestFixture;
 import cart.dto.ResponseFixture;
 import org.junit.jupiter.api.Test;
