+++ src/test/java/cart/controller/api/ProductRequestControllerTest.java
@@ -4,7 +4,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import cart.controller.AbstractProductControllerTest;
-import cart.dto.ProductRequest;
+import cart.controller.dto.ProductRequest;
 import org.junit.jupiter.api.Test;
 
 @SuppressWarnings({"NonAsciiCharacters", "SpellCheckingInspection"})
