+++ src/test/java/cart/controller/view/UserViewControllerTest.java
@@ -10,8 +10,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import java.util.List;
 
 import cart.controller.AbstractProductControllerTest;
+import cart.controller.dto.UserResponse;
 import cart.domain.user.User;
-import cart.dto.UserResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
