+++ src/test/java/cart/dto/ResponseFixture.java
@@ -1,5 +1,7 @@
 package cart.dto;
 
+import cart.controller.dto.ProductResponse;
+
 @SuppressWarnings("SpellCheckingInspection")
 public class ResponseFixture {
