+++ src/test/java/cart/repository/StubCartRepository.java
@@ -6,9 +6,9 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 
+import cart.controller.dto.CartResponse;
 import cart.domain.cart.Cart;
 import cart.domain.user.Email;
-import cart.dto.CartResponse;
 import cart.entiy.CartEntity;
 
 public class StubCartRepository implements CartRepository {
